[CMA-10] refactor : 패키지 구조 수정
- CreateGoalFIrstPageFragment 패키지 구조 수정
